removed proxy-forward and chain-web-requests, added web-client-get-page
the processors were removed because their quality was not up to the expected standard.
if they were to be included again, they should be rewritten to not use classes
from either piped or twisted test suites.

in the meanwhile, web-client-get-page supports much of the functionality
previously supported by proxy-forward, with the exception of getting received
headers/cookies.